requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Deprecated==1.2.13
filelock==3.4.0
Flask==2.0.2
Flask-RESTful==0.3.9
flask-talisman==0.8.1
huggingface-hub==0.2.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.5
packaging==21.3
pymongo==4.1.1
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
redis==4.3.3
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
six==1.16.0
tokenizers==0.10.3
torch==1.10.1
tqdm==4.62.3
transformers==4.15.0
typing_extensions==4.0.1
urllib3==1.26.7
waitress==2.0.0
Werkzeug==2.0.2
wrapt==1.14.1
celery~=5.2.7